// WFX_Code/WFXWeb/WFX.Web/ClientApp/src/models/POPlanTargetModel.ts
export class POPlanTargetModel {
    
    soNo: string;
    poNo: string;
    style: string;
    StartDate: string;
    EndDate: string;
    ProcessCode: string;

    solist: {
      id: string;
      text: string;
    }
    polist: {
      id: string;
      text: string;
    }
    processlist:{
      id: string;
      text: string;
    }
    
  }

  